refactore types and utils for teamuser
