+++ src/server/api/routers/teamUser/teamUser-types.ts
@@ -1,0 +1,5 @@
+import { type TeamUser, type LeagueUser } from "@prisma/client";
+export interface TeamUserAndLeagueUser {
+  teamUser: TeamUser;
+  leagueUser: LeagueUser;
+}
